refactor to split code into multiple files
